+++ core/model/src/main/java/dev/vitalcc/stukay/core/model/Identifiers.kt
@@ -1,0 +1,13 @@
+package dev.vitalcc.stukay.core.model
+
+@JvmInline
+value class ProjectId(val value: String)
+
+@JvmInline
+value class ThreadId(val value: String)
+
+@JvmInline
+value class TurnId(val value: String)
+
+@JvmInline
+value class ApprovalId(val value: String)